Enhance G-Purchasing

+++ src/modules/garment-purchasing/unit-expenditure-note-all/view.js
@@ -96,17 +96,26 @@ export class View {
     }
 
     cancel(event) {
-        this.router.navigateToRoute('list');
+        var r = confirm("Apakah anda yakin akan keluar?")
+        if ( r == true) {
+            this.router.navigateToRoute('list');
+        }
     }
 
     edit(event) {
-        this.router.navigateToRoute('edit', { id: this.data.Id });
+        var r = confirm("Apakah anda yakin akan mengubah data ini?")
+        if (r == true) {
+            this.router.navigateToRoute('edit', { id: this.data.Id });
+        }
     }
 
     delete(event) {
-        this.service.delete(this.data).then(result => {
-            this.cancel();
-        });
+        var r = confirm("Apakah anda yakin akan menghapus data ini?")
+        if (r == true) {
+            this.service.delete(this.data).then(result => {
+                this.cancel();
+            });
+        }
     }
 
 }
